+++ server/model/updateNote.go
@@ -2,7 +2,7 @@ package model
 
 // UpdateNote struct for update notes
 type UpdateNote struct {
-	VersionCode int32  `gorm:"primary_key;AUTO_INCREMENT;column:version_code;type:int;"`
+	VersionCode int64  `gorm:"primary_key;AUTO_INCREMENT;column:version_code;type:int;"`
 	VersionName string `gorm:"column:version_name;type:text;"`
 	Message     string `gorm:"column:message;type:text;"`
 }
